+++ apps/server/src/modules/account/services/dto/index.ts
@@ -1,2 +1,2 @@
-export * from './account.dto';
+export * from './account-read.dto';
 export * from './account-save.dto';
